fix(refactor): Refactor code as per review to simplify ClassDependenciesGraph
